simplify the logic that judge whether a file is a symlink or not
correct the method that determine the real path of symlinks

add extra comment

+++ functions.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 function timestamp()
 {
 	stat -L -c '%Y' $1
@@ -5,87 +6,104 @@ function timestamp()
 
 function atomic_cp()
 {
-	dstdir=$(dirname $2)
+	dstdir="$(dirname $2)"
 	if [ ! -d "$dstdir" ]; then
-		mkdir -p $dstdir
+		mkdir -p "$dstdir"
 	fi
 
-	tmpfile=$(mktemp --tmpdir=$CACHETMPDIR)
-    #####
-	#cp -L --preserve=all $1 $tmpfile
-	#mv $tmpfile $2
-    #####
-
-    
     ##### Explanation
-    # 1. find out the *REAL* path of $dstfile and store it in $oriabspath
-    # 2. if original file is a link, then
-    #        if the file that is pointed to has not been cached
-    #            1. copy the original file and the dirs that contains it to ssd cache
-    #            2. create a symbolic link in $dstfile and let it points to the cached original file
-    #        if the file has been cached
-    #            1. just create the link
-    #    if original file is an ordinary file
-    #        just copy
+    # 1. $oriabspath stands for the *REAL* path of $dstfile
+    # 2. if original file is a link
+    #        if the file that is pointed to hasn't been cached
+    #            copy the original file and the dirs that contains it to ssd-cache
+    #            copy the original link(path should be relative) to destination
+    #        elif the file has been cached
+    #            create the link
+    #    elif original file is an ordinary file
+    #        copy
     #####
 
-    #####
-    dstfile=$2
+    if [ -L "$1" ]; then
+        dstfile="${2#$CACHEROOT}"   # e.g $dstfile = /ubuntu-releases/14.04.2/ubuntu-14.04.2-desktop-amd64.iso
+        t="${dstfile#/*/}"          # e.g $t = 14.04.2/ubuntu-14.04.2-desktop-amd64.iso
+        repo="${dstfile%$t}"        # e.g $repo = /ubuntu-releases/
+        dstfile="$t"
 
-    oriabspath="$(readlink -e $dstfile)"
-    [ -z $oriabspath ] && return  # broken link or file not exist. In genernal, should never get here
+        local oriabspath="$(readlink -e $1)" # e.g $oriabspath = /mnt/repo/ubuntu-releases/.pool/ubuntu-14.04.2-desktop-amd64.iso
+        [[ -z $oriabspath ]] && return
 
-    ORIGINROOT="${oriabspath%$dstfile}"
+        ORIGINROOT="${oriabspath%%$repo*}"
+        ORIGINROOT+="$repo"         # e.g $ORIGINROOT = /mnt/repo/ubuntu-releases/
 
-    if [ $oriabspath != $ORIGINROOT$dstfile ]; then  # original file is a link
-        linktoname="${oriabspath#$ORIGINROOT}"
-        if [ ! -f $CACHEROOT$linktoname ]; then
-            linktodir="$(dirname $CACHEROOT$linktoname)"
+        linktoname="${oriabspath#$ORIGINROOT}"   # e.g $linktoname = .pool/ubuntu-14.04.2-desktop-amd64.iso
+        if [ ! -f $CACHEROOT$repo$linktoname ]; then
+
+            if [[ $linktoname == *mnt* ]]; then
+                echo "-------------Wrong Path---------------" >>/tmp/wrongpath.log
+                echo "linktoname: $linktoname" >>/tmp/wrongpath.log
+                echo "source: $1" >>/tmp/wrongpath.log
+                echo "destination: $2" >>/tmp/wrongpath.log
+                echo "dstfile: $dstfile" >>/tmp/wrongpath.log
+                echo "repo: $repo" >>/tmp/wrongpath.log
+                echo "oriabspath: $oriabspath" >>/tmp/wrongpath.log
+                echo "ORIGINROOT: $ORIGINROOT" >>/tmp/wrongpath.log
+                return
+            fi
+
+            linktodir="$(dirname $CACHEROOT$repo$linktoname)"
             [ ! -d "$linktodir" ] && mkdir -p "$linktodir"
-            cp --preserve=all $oriabspath $tmpfile
-            mv $tmpfile $dstfile
-        fi
 
-        # add --force. in case that the link has been modified to point to other place
-        ln -sf $CACHEROOT$linktoname $CACHEROOT$dstfile 
+            local tmpfile="$(mktemp --tmpdir=$CACHETMPDIR)"
+            #local tmpfile="$(mktemp -u --tmpdir=$CACHETMPDIR)"
+            cp --preserve=all $oriabspath $tmpfile
+            mv $tmpfile $CACHEROOT$repo$linktoname
+            #echo "cp --preserve=all $oriabspath $tmpfile"
+            #echo "mv $tmpfile $CACHEROOT$repo$linktoname"
+        fi
+        ln -sf $CACHEROOT$repo$linktoname $2
+        #echo ln -sf $CACHEROOT$repo$linktoname $2
+
     else
-        cp --preserve=all $oriabspath $tmpfile
-        mv $tmpfile $dstfile
+        local tmpfile="$(mktemp --tmpdir=$CACHETMPDIR)"
+        #local tmpfile="$(mktemp -u --tmpdir=$CACHETMPDIR)"
+        cp --preserve=all $1 $tmpfile
+        mv $tmpfile $2
+        #echo "cp --preserve=all $1 $tmpfile"
+        #echo "mv $tmpfile $2"
     fi
-    #####
 }
 
 function remove_uncached_files()
 {
+    local tmpfile="$(mktemp --tmpdir=$CACHETMPDIR keep.XXXXXXXXXX)"
+    cp $1 $tmpfile
+    cat $1 | while read f; do
+        if [ -L "$CACHEROOT$f" ]; then
+            local oriabspath="$(readlink -e $CACHEROOT$f)"
+            [ -n "$oriabspath" ] && echo "${oriabspath#$CACHEROOT}" >> $tmpfile
+        fi
+    done
+
 	diff --old-line-format='' --new-line-format='%L' --unchanged-line-format='' \
-		<(cat $1 | sort | uniq) \
-		<(find $CACHEROOT -type f | cut -c $((${#CACHEROOT}+1))- | sort | uniq) \
+		<(cat $tmpfile | sort -u) \
+		<(find $CACHEROOT -type f -o -type l | cut -c $((${#CACHEROOT}+1))- | sort -u) \
 	| while read f; do
+
 		echo $f
-
-        #####
-        # No matter it is a real file or a link, just remove it.
-        # try_files in nginx will handle the error case properly.
-        #####
-
-		rm -rf $CACHEROOT$f
+		rm -rf "$CACHEROOT$f"
 	done
+    #rm $tmpfile
 }
 function remove_expired_files()
 {
-	find $CACHEROOT -type f | cut -c $((${#CACHEROOT}+1))- \
+	find $CACHEROOT -type f -o -type l | cut -c $((${#CACHEROOT}+1))- \
 	| while read f; do
 		if [ -f "$WWWROOT$f" ] && [ "$(timestamp $WWWROOT$f)" == "$(timestamp $CACHEROOT$f)" ]; then
 			continue
 		fi
+
 		echo $f
-
-        #####
-        # No matter it is a real file or a link, just remove it.
-        # try_files in nginx will handle the error case properly.
-        #####
-
-		rm -rf $CACHEROOT$f
+		rm -rf "$CACHEROOT$f"
 	done
 }
 function echo_timestamp()
@@ -139,20 +157,21 @@ function sync_from_file_list()
 
 	echo_timestamp
 
-    #####
 	echo "===== Removing broken links ====="
-    find $CACHEROOT -type l -xtype l -delete
-    #####
+    find $CACHEROOT -type l -xtype l -print -delete
+
+	echo_timestamp
 
 	echo "===== Synchronizing new files ====="
 	cat $cache_list | while read f; do
-		[ ! -f "$WWWROOT$f" ] && continue # source file not exist or is a directory
+        # source file not exist or is a directory
+		[ ! -f "$WWWROOT$f" ] && continue
 
-		if [ -d "$CACHEROOT$f" ]; then # it was a directory, now a file
-			rm -rf $CACHEROOT$f
-		fi
+        # it was a directory, now a file
+		[ -d "$CACHEROOT$f" ] && rm -rf $CACHEROOT$f
 
-		if [ ! -f "$CACHEROOT$f" ]; then # file not cached
+        # file not cached
+		if [ ! -f "$CACHEROOT$f" ]; then
 			echo $f
 			atomic_cp $WWWROOT$f $CACHEROOT$f
 		fi
